requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==25.1.0
audioop-lts==0.2.1
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
discord==2.3.2
discord.py==2.4.0
frozenlist==1.5.0
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
oauthlib==3.2.2
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
py-cord==2.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
pytest==8.3.4
pytest-mock==3.14.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3
